+++ src/core/hle/service/hid/controllers/xpad.cpp
@@ -20,7 +20,7 @@ void Controller_XPad::OnRelease() {}
 void Controller_XPad::OnUpdate(const Core::Timing::CoreTiming& core_timing, u8* data,
                                std::size_t size) {
     for (auto& xpad_entry : shared_memory.shared_memory_entries) {
-        xpad_entry.header.timestamp = core_timing.GetCPUTicks();
+        xpad_entry.header.timestamp = static_cast<s64>(core_timing.GetCPUTicks());
         xpad_entry.header.total_entry_count = 17;
 
         if (!IsControllerActivated()) {
@@ -30,9 +30,11 @@ void Controller_XPad::OnUpdate(const Core::Timing::CoreTiming& core_timing, u8* 
         }
         xpad_entry.header.entry_count = 16;
 
-        const auto& last_entry = xpad_entry.pad_states[xpad_entry.header.last_entry_index];
+        const auto& last_entry =
+            xpad_entry.pad_states[static_cast<u64>(xpad_entry.header.last_entry_index)];
         xpad_entry.header.last_entry_index = (xpad_entry.header.last_entry_index + 1) % 17;
-        auto& cur_entry = xpad_entry.pad_states[xpad_entry.header.last_entry_index];
+        auto& cur_entry =
+            xpad_entry.pad_states[static_cast<u64>(xpad_entry.header.last_entry_index)];
 
         cur_entry.sampling_number = last_entry.sampling_number + 1;
         cur_entry.sampling_number2 = cur_entry.sampling_number;
