+++ src/core/network/network.h
@@ -67,12 +67,12 @@ struct PollFD {
     u16 revents;
 };
 
-constexpr u16 POLL_IN = 1 << 0;
-constexpr u16 POLL_PRI = 1 << 1;
-constexpr u16 POLL_OUT = 1 << 2;
-constexpr u16 POLL_ERR = 1 << 3;
-constexpr u16 POLL_HUP = 1 << 4;
-constexpr u16 POLL_NVAL = 1 << 5;
+constexpr u32 POLL_IN = 1 << 0;
+constexpr u32 POLL_PRI = 1 << 1;
+constexpr u32 POLL_OUT = 1 << 2;
+constexpr u32 POLL_ERR = 1 << 3;
+constexpr u32 POLL_HUP = 1 << 4;
+constexpr u32 POLL_NVAL = 1 << 5;
 
 class NetworkInstance {
 public:
